third_party/WebKit/LayoutTests/editing/inserting/insert-div-009.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>abc|</div>',
        '</div>',
    ].join(''),
    'insertParagraph',
    [
        '<div contenteditable>',
            '<div>abc</div>',
            '<div>|<br></div>',
        '</div>',
    ].join('')),
    'insert BR at end of block');
</script>
